+++ src/components/layer-toggle/checkbox-type/styles.module.scss
